requirements.txt
# requirements.txt
pandas
streamlit
google-genai
numpy
plotly # For potential future visualizations
